Award Management

+++ src/main/java/com/example/demo/repository/AwardRepository.java
@@ -3,4 +3,4 @@ package com.example.demo.repository;
 import com.example.demo.model.Award;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AwardRepository extends JpaRepository<Award, Long> {}
+public interface AwardRepository extends JpaRepository<Award, Integer> {}
